Refactor dbaccess to dataaccesslayer
